futures/requirements.txt
alpaca-py >= 0.43.2
pandas >= 2.0.0
numpy >= 1.24.0
python-dotenv >= 1.0.0
pytz >= 2024.1
matplotlib >= 3.7.0
ib_insync >= 0.9.86
